Car-Pooling-App/user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
# psycopg2
pika
# pydantic
python-jose[cryptography]
passlib[bcrypt]
email-validator
